Verify compatibility of provided Test262 directory
Future versions of that project may alter the requirements for test
generation. Because the user of this library may provides any arbitrary
version of Test262, it is possible that some future change would cause
this library to emit invalid test material.

Reference the `package.json` file provided by Test262 to determine
compatibility and reject incompatible versions. Because consumers may
only be interested in a subset of tests which are not effected by any
given "breaking change," expose a configuration option named
`acceptVersoin` to disable this restriction.