Include `test262-compiler`
The functionality previously delegated to the externally-maintained
`test262-compiler` module are core to this module's behavior. Include
that source code in order to better facilitate maintenance.

The `test262-compiler` module exposes a command-line interface uses
the`yargs` module. Because that feature is unnecessary in this context,
this change also reduces the number of dependencies from 109 to 62.