Simplify internals
Previously, test "compilation" was performed by an external module. This
project included code to manipulate the output of that module in a
sub-optimal way. Because a recent commit incorporated the external
module into this project directly, the sub-optimal code is no longer
necessary.

Refactor the internals to produce the same output in a more readable and
efficient way.