Remove undocumented modification
This project inherited code from the `test262-compiler` project which
was tightly coupled to the `test262-harness` project. This coupling is
apparent in a modification made to synchronous tests: the insertion of
an invocation of the `$DONE` function. Test262's interpreting guidelines
make no mention of such a transformation.

Remove the modification in order to comply with that document.
(Consumers of `test262-stream` who rely on this functionality may
re-implement it by transforming the test stream.)