Update JSHint
Prior releases of JSHint included an insecure version of the Lodash npm
module [1]. The latest release of JSHint depends on a more recent
version of Lodash which does not suffer from the vulnerability.

Although the risk from a development dependency like JSHint is low, some
tooling (e.g. GitHub.com and npm) reports security warnings which may
alarm contributors. Update to the latest release in order to avoid
security warnings.

[1] https://nvd.nist.gov/vuln/detail/CVE-2018-3721